Update withdraw page to also support logging in
